requirements.txt
Flask
Flask_SQLAlchemy
gunicorn
pymysql
cryptography
python-dotenv
SQLAlchemy